TaskLauncherActor doesn't wait for in-flight tasks on stop
Summary:
I don't see a particular reason for the TaskLauncherActor to wait 30s for updates
on inflight tasks if it's being stopped. So I decided to remove it and see what happens.
There is a corresponding jenkins loop here: https://jenkins.mesosphere.com/service/jenkins/view/Marathon/job/marathon-sandbox/job/marathon-loop-ad-fix-group-deploy-it/

If you know why it was there in the first place - let me know.

Test Plan: unit, it

Reviewers: jdef, timcharper, jeschkies, unterstein, meichstedt

Subscribers: marathon-team, marathon-dev, jenkins

Differential Revision: https://phabricator.mesosphere.com/D907