Workaround for flaky CleanUp in ResidentTaskIT
Summary:
Gist of the problem is that the scale check and the clean up happen at the same time
leading to a problem. I documented the detailed findings in MARATHON-7338. A simple workaround
can be just letting the task run longer avoiding the situation altogether.

Test Plan: it

Reviewers: jeschkies, meichstedt

Subscribers: marathon-dev, jenkins, marathon-team

JIRA Issues: MARATHON-7338

Differential Revision: https://phabricator.mesosphere.com/D998